refactor: 의존성에 따른 위치 조정
