+++ src/js/domain/models/entities/Lotto/Lotto.js
@@ -1,4 +1,4 @@
-import { hasDuplicated } from '../../../utils/utils.js';
+import { hasDuplicated } from '../../../../utils.js';
 import {
     NumbersNotArrayError,
     NumbersLengthNotSixError,
