+++ src/js/domain/models/entities/LottoStore/errors.js
@@ -9,25 +9,25 @@ class LottoStoreError extends DomainError {
 }
 
 export class BudgetNotNumberError extends LottoStoreError {
-    static MESSAGE = 'budget은 Number 타입이어야합니다.';
+    static #MESSAGE = 'budget은 Number 타입이어야합니다.';
 
     constructor() {
-        super(BudgetNotNumberError.MESSAGE);
+        super(BudgetNotNumberError.#MESSAGE);
     }
 }
 
 export class BudgetBelowMinError extends LottoStoreError {
-    static MESSAGE = 'budget은 1_000 이상의 값이어야합니다.';
+    static #MESSAGE = 'budget은 1_000 이상의 값이어야합니다.';
 
     constructor() {
-        super(BudgetBelowMinError.MESSAGE);
+        super(BudgetBelowMinError.#MESSAGE);
     }
 }
 
 export class BudgetAboveMaxError extends LottoStoreError {
-    static MESSAGE = 'budget은 100_000 이하의 값이어야합니다.';
+    static #MESSAGE = 'budget은 100_000 이하의 값이어야합니다.';
 
     constructor() {
-        super(BudgetAboveMaxError.MESSAGE);
+        super(BudgetAboveMaxError.#MESSAGE);
     }
 }
